Split up canvas spec, extract community nodes related tests into its own spec, various flakiness fixes
