+++ cypress/e2e/12-canvas.cy.ts
@@ -20,7 +20,7 @@ const ZOOM_OUT_X1_FACTOR = 0.8;
 const ZOOM_OUT_X2_FACTOR = 0.64;
 const RENAME_NODE_NAME = 'Something else';
 
-describe('Canvas Actions', () => {
+describe('Canvas Node Manipulation and Navigation', () => {
 	before(() => {
 		cy.resetAll();
 		cy.skipSetup();
@@ -30,56 +30,6 @@ describe('Canvas Actions', () => {
 		WorkflowPage.actions.visit();
 	});
 
-	it('should render canvas', () => {
-		WorkflowPage.getters.nodeViewRoot().should('be.visible');
-		WorkflowPage.getters.canvasPlusButton().should('be.visible');
-		WorkflowPage.getters.zoomToFitButton().should('be.visible');
-		WorkflowPage.getters.zoomInButton().should('be.visible');
-		WorkflowPage.getters.zoomOutButton().should('be.visible');
-		WorkflowPage.getters.executeWorkflowButton().should('be.visible');
-	});
-
-	it('should connect and disconnect a simple node', () => {
-		WorkflowPage.actions.addNodeToCanvas(SET_NODE_NAME);
-		WorkflowPage.getters.nodeViewBackground().click(600, 200, { force: true });
-		cy.get('.jtk-connector').should('have.length', 1);
-		WorkflowPage.actions.addNodeToCanvas(SET_NODE_NAME);
-
-		// Change connection from Set to Set1
-		cy.draganddrop(
-			WorkflowPage.getters.getEndpointSelector('input', SET_NODE_NAME),
-			WorkflowPage.getters.getEndpointSelector('input', `${SET_NODE_NAME}1`),
-		);
-
-		WorkflowPage.getters
-			.canvasNodeInputEndpointByName(`${SET_NODE_NAME}1`)
-			.should('have.class', 'jtk-endpoint-connected');
-
-		cy.get('.jtk-connector').should('have.length', 1);
-		// Disconnect Set1
-		cy.drag(WorkflowPage.getters.getEndpointSelector('input', `${SET_NODE_NAME}1`), [-200, 100]);
-		cy.get('.jtk-connector').should('have.length', 0);
-	});
-
-	it('should add first step', () => {
-		WorkflowPage.getters.canvasPlusButton().should('be.visible');
-		WorkflowPage.actions.addInitialNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.getters.canvasNodes().should('have.length', 1);
-	});
-
-	it('should add a node via plus endpoint drag', () => {
-		WorkflowPage.getters.canvasPlusButton().should('be.visible');
-		WorkflowPage.actions.addNodeToCanvas(SCHEDULE_TRIGGER_NODE_NAME, true);
-
-		cy.drag(
-			WorkflowPage.getters.getEndpointSelector('plus', SCHEDULE_TRIGGER_NODE_NAME),
-			[100, 100],
-		);
-
-		WorkflowPage.getters.nodeCreatorSearchBar().should('be.visible');
-		WorkflowPage.actions.addNodeToCanvas(IF_NODE_NAME, false);
-		WorkflowPage.getters.nodeViewBackground().click({ force: true });
-	});
 
 	it('should add switch node and test connections', () => {
 		WorkflowPage.actions.addNodeToCanvas(SWITCH_NODE_NAME, true);
@@ -142,6 +92,8 @@ describe('Canvas Actions', () => {
 
 		cy.get('.rect-input-endpoint.jtk-endpoint-connected').should('have.length', 4);
 		WorkflowPage.actions.executeWorkflow();
+		WorkflowPage.getters.stopExecutionButton().should('not.exist');
+
 		// If the merged set nodes are connected and executed correctly, there should be 2 items in the output of merge node
 		cy.get('[data-label="2 items"]').should('be.visible');
 	});
@@ -165,41 +117,6 @@ describe('Canvas Actions', () => {
 		cy.get('.plus-draggable-endpoint').filter(':visible').should('not.have.class', 'ep-success');
 		cy.get('.jtk-connector.success').should('have.length', 3);
 		cy.get('.jtk-connector').should('have.length', 4);
-	});
-
-	it('should add a connected node using plus endpoint', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		cy.get('.plus-endpoint').should('be.visible').click();
-		WorkflowPage.getters.nodeCreatorSearchBar().should('be.visible');
-		WorkflowPage.getters.nodeCreatorSearchBar().type(CODE_NODE_NAME);
-		WorkflowPage.getters.nodeCreatorSearchBar().type('{enter}');
-		cy.get('body').type('{esc}');
-		WorkflowPage.getters.canvasNodes().should('have.length', 2);
-		WorkflowPage.getters.nodeConnections().should('have.length', 1);
-	});
-
-	it('should add disconnected node if nothing is selected', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		// Deselect nodes
-		WorkflowPage.getters.nodeViewBackground().click({ force: true });
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		WorkflowPage.getters.canvasNodes().should('have.length', 2);
-		WorkflowPage.getters.nodeConnections().should('have.length', 0);
-	});
-
-	it('should add note between two connected nodes', () => {
-		WorkflowPage.actions.addNodeToCanvas(SCHEDULE_TRIGGER_NODE_NAME);
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		WorkflowPage.actions.addNodeToCanvas(SET_NODE_NAME);
-		WorkflowPage.actions.zoomToFit();
-		WorkflowPage.actions.addNodeBetweenNodes(CODE_NODE_NAME, SET_NODE_NAME, HTTP_REQUEST_NODE_NAME);
-		WorkflowPage.getters.canvasNodes().should('have.length', 4);
-		WorkflowPage.getters.nodeConnections().should('have.length', 3);
-		// And last node should be pushed to the right
-		WorkflowPage.getters
-			.canvasNodes()
-			.last()
-			.should('have.attr', 'style', 'left: 860px; top: 260px;');
 	});
 
 	it('should delete node using node action button', () => {
@@ -319,50 +236,6 @@ describe('Canvas Actions', () => {
 		WorkflowPage.getters.canvasNodes().last().should('be.visible');
 	});
 
-	it('should select all nodes', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		WorkflowPage.actions.selectAll();
-		WorkflowPage.getters.selectedNodes().should('have.length', 2);
-	});
-
-	it('should select nodes using arrow keys', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.getters.canvasNodeByName(MANUAL_TRIGGER_NODE_DISPLAY_NAME).click();
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		cy.wait(500);
-		cy.get('body').type('{leftArrow}');
-		WorkflowPage.getters.canvasNodes().first().should('have.class', 'jtk-drag-selected');
-		cy.get('body').type('{rightArrow}');
-		WorkflowPage.getters.canvasNodes().last().should('have.class', 'jtk-drag-selected');
-	});
-
-	it('should select nodes using shift and arrow keys', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.getters.canvasNodeByName(MANUAL_TRIGGER_NODE_DISPLAY_NAME).click();
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		cy.wait(500);
-		cy.get('body').type('{shift}', { release: false }).type('{leftArrow}');
-		WorkflowPage.getters.selectedNodes().should('have.length', 2);
-	});
-
-	it('should delete connections by pressing the delete button', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.getters.canvasNodeByName(MANUAL_TRIGGER_NODE_DISPLAY_NAME).click();
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		WorkflowPage.getters.nodeConnections().first().realHover();
-		cy.get('.connection-actions .delete').first().click({ force: true });
-		WorkflowPage.getters.nodeConnections().should('have.length', 0);
-	});
-
-	it('should delete a connection by moving it away from endpoint', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.getters.canvasNodeByName(MANUAL_TRIGGER_NODE_DISPLAY_NAME).click();
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		cy.drag(WorkflowPage.getters.getEndpointSelector('input', CODE_NODE_NAME), [0, -100]);
-		WorkflowPage.getters.nodeConnections().should('have.length', 0);
-	});
-
 	it('should disable node by pressing the disable button', () => {
 		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
 		WorkflowPage.getters.canvasNodeByName(MANUAL_TRIGGER_NODE_DISPLAY_NAME).click();
@@ -417,24 +290,5 @@ describe('Canvas Actions', () => {
 			.click({ force: true });
 		WorkflowPage.getters.canvasNodes().should('have.length', 3);
 		WorkflowPage.getters.nodeConnections().should('have.length', 1);
-	});
-
-	it('should execute node', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		WorkflowPage.getters
-			.canvasNodes()
-			.last()
-			.find('[data-test-id="execute-node-button"]')
-			.click({ force: true });
-		WorkflowPage.getters.successToast().should('contain', 'Node executed successfully');
-	});
-
-	it('should copy selected nodes', () => {
-		WorkflowPage.actions.addNodeToCanvas(MANUAL_TRIGGER_NODE_NAME);
-		WorkflowPage.actions.addNodeToCanvas(CODE_NODE_NAME);
-		WorkflowPage.actions.selectAll();
-		WorkflowPage.actions.hitCopy();
-		WorkflowPage.getters.successToast().should('contain', 'Copied!');
 	});
 });
