+++ cypress/e2e/2-credentials.cy.ts
@@ -16,6 +16,9 @@ import {
 } from '../constants';
 import { randFirstName, randLastName } from '@ngneat/falso';
 import { CredentialsPage, CredentialsModal, WorkflowPage, NDV } from '../pages';
+import CustomNodeWithN8nCredentialFixture from '../fixtures/Custom_node_n8n_credential.json';
+import CustomNodeWithCustomCredentialFixture from '../fixtures/Custom_node_custom_credential.json';
+import CustomCredential from '../fixtures/Custom_credential.json';
 
 const email = DEFAULT_USER_EMAIL;
 const password = DEFAULT_USER_PASSWORD;
@@ -31,25 +34,10 @@ const NEW_CREDENTIAL_NAME = 'Something else';
 describe('Credentials', () => {
 	before(() => {
 		cy.resetAll();
-		cy.setup({ email, firstName, lastName, password });
-
-		// Always intercept the request to test credentials and return a success
-		cy.intercept('POST', '/rest/credentials/test', {
-			statusCode: 200,
-			body: {
-				data: { status: 'success', message: 'Tested successfully' },
-			}
-		});
+		cy.skipSetup();
 	});
 
 	beforeEach(() => {
-		cy.on('uncaught:exception', (err, runnable) => {
-			expect(err.message).to.include('Not logged in');
-
-			return false;
-		});
-
-		cy.signin({ email, password });
 		cy.visit(credentialsPage.url);
 	});
 
@@ -249,25 +237,5 @@ describe('Credentials', () => {
 		workflowPage.getters.nodeCredentialsSelect().find('li').last().click();
 		credentialsModal.actions.fillCredentialsForm();
 		workflowPage.getters.nodeCredentialsSelect().should('contain', NEW_QUERY_AUTH_ACCOUNT_NAME);
-	});
-
-	it('should render custom node with n8n credential', () => {
-		workflowPage.actions.visit();
-		workflowPage.actions.addNodeToCanvas('Manual');
-		workflowPage.actions.addNodeToCanvas('E2E Node with native n8n credential', true, true);
-		workflowPage.getters.nodeCredentialsLabel().click();
-		cy.contains('Create New Credential').click();
-		credentialsModal.getters.editCredentialModal().should('be.visible');
-		credentialsModal.getters.editCredentialModal().should('contain.text', 'Notion API');
-	});
-
-	it('should render custom node with custom credential', () => {
-		workflowPage.actions.visit();
-		workflowPage.actions.addNodeToCanvas('Manual');
-		workflowPage.actions.addNodeToCanvas('E2E Node with custom credential', true, true);
-		workflowPage.getters.nodeCredentialsLabel().click();
-		cy.contains('Create New Credential').click();
-		credentialsModal.getters.editCredentialModal().should('be.visible');
-		credentialsModal.getters.editCredentialModal().should('contain.text', 'Custom E2E Credential');
 	});
 });
